requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bert-score==0.3.13
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
distro==1.9.0
evaluate==0.4.4
executing==2.2.0
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2024.12.0
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-experimental==0.3.4
langchain-neo4j==0.4.0
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langsmith==0.4.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
mypy_extensions==1.1.0
neo4j==5.28.1
neo4j-graphrag==1.7.0
networkx==3.5
nltk==3.9.1
numpy==2.3.0
openai==1.90.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rouge_score==0.1.2
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
types-PyYAML==6.0.12.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.1
yfiles_jupyter_graphs==1.10.7
zstandard==0.23.0
